+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/compte/page.tsx
@@ -1,8 +1,13 @@
+import type { Metadata } from "next"
 import { ADMIN } from "shared/constants/recruteur"
 
 import CompteRenderer from "@/app/(espace-pro)/espace-pro/(connected)/_components/CompteRenderer"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import { PAGES } from "@/utils/routes.utils"
+
+export const metadata: Metadata = {
+  title: PAGES.dynamic.compte({ userType: ADMIN }).getMetadata().title,
+}
 
 export default function Page() {
   return (
