+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/gestion-des-entreprises/page.tsx
@@ -1,4 +1,10 @@
+import type { Metadata } from "next"
 import GestionEntreprises from "./gestionEntreprises"
+import { PAGES } from "@/utils/routes.utils"
+
+export const metadata: Metadata = {
+  title: PAGES.static.backAdminGestionDesEntreprises.getMetadata().title,
+}
 
 export default async function AdministrationGestionEntreprises() {
   return <GestionEntreprises />
