+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/processeur/ProcesseurPage.tsx
@@ -1,0 +1,22 @@
+"use client"
+import { Box, Typography } from "@mui/material"
+import { ProcessorStatusIndexComponent } from "job-processor/dist/react"
+
+import { ProcessorStatusProvider } from "./components/ProcessorStatusProvider"
+import { Breadcrumb } from "@/app/_components/Breadcrumb"
+import { publicConfig } from "@/config.public"
+import { PAGES } from "@/utils/routes.utils"
+
+export default function AdminProcessorPage() {
+  return (
+    <Box>
+      <Breadcrumb pages={[PAGES.static.adminProcessor]} />
+      <Typography variant="h2" gutterBottom>
+        {PAGES.static.adminProcessor.title}
+      </Typography>
+      <ProcessorStatusProvider>
+        {(status) => <ProcessorStatusIndexComponent status={status} baseUrl={new URL(PAGES.static.adminProcessor.getPath(), publicConfig.baseUrl).href} />}
+      </ProcessorStatusProvider>
+    </Box>
+  )
+}
