+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/processeur/job/[name]/page.tsx
@@ -1,25 +1,14 @@
-"use client"
-import { Box, Typography } from "@mui/material"
-import { ProcessorStatusJobComponent } from "job-processor/dist/react"
-import { use } from "react"
-
-import { ProcessorStatusProvider } from "@/app/(espace-pro)/espace-pro/(connected)/administration/processeur/components/ProcessorStatusProvider"
-import { Breadcrumb } from "@/app/_components/Breadcrumb"
-import { publicConfig } from "@/config.public"
+import type { Metadata } from "next"
+import ProcesseurJobPage from "./ProcesseurJobPage"
 import { PAGES } from "@/utils/routes.utils"
 
-export default function JobTypePage({ params }: { params: Promise<{ name: string }> }) {
-  const { name: rawName } = use(params)
-  const name = decodeURIComponent(rawName)
-  return (
-    <Box>
-      <Breadcrumb pages={[PAGES.static.adminProcessor, PAGES.dynamic.adminProcessorJob(name)]} />
-      <Typography variant="h2" gutterBottom>
-        {PAGES.dynamic.adminProcessorJob(name).title}
-      </Typography>
-      <ProcessorStatusProvider>
-        {(status) => <ProcessorStatusJobComponent name={name} status={status} baseUrl={new URL(PAGES.static.adminProcessor.getPath(), publicConfig.baseUrl).href} />}
-      </ProcessorStatusProvider>
-    </Box>
-  )
+export async function generateMetadata({ params }: { params: Promise<{ name: string }> }): Promise<Metadata> {
+  const { name } = await params
+  return {
+    title: PAGES.dynamic.adminProcessorJob(name).getMetadata().title,
+  }
+}
+
+export default async function Page() {
+  return <ProcesseurJobPage />
 }
