+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/rendez-vous-apprentissage/[siret]/page.tsx
@@ -1,6 +1,15 @@
+import type { Metadata } from "next"
 import RendezVousApprentissageDetailRendererClient from "./RendezVousApprentissageDetailRendererClient"
 import LoadingEmptySpace from "@/app/(espace-pro)/_components/LoadingEmptySpace"
 import { getEligibleTrainingsForAppointments, getEtablissement } from "@/utils/api"
+import { PAGES } from "@/utils/routes.utils"
+
+export async function generateMetadata({ params }: { params: Promise<{ siret: string }> }): Promise<Metadata> {
+  const { siret } = await params
+  return {
+    title: PAGES.dynamic.rendezVousApprentissageDetail({ siret }).getMetadata().title,
+  }
+}
 
 export default async function RendezVousApprentissageDetail({ params }: { params: Promise<{ siret: string }> }) {
   const { siret } = await params
