+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/users/[userId]/entreprise/[establishment_id]/offre/[job_id]/page.tsx
@@ -1,35 +1,16 @@
-"use client"
-
-import { useParams, useRouter } from "next/navigation"
+import type { Metadata } from "next"
 import { ADMIN } from "shared/constants/recruteur"
 
-import UpsertOffre from "@/app/(espace-pro)/espace-pro/(connected)/_components/UpsertOffre"
-import { Breadcrumb } from "@/app/_components/Breadcrumb"
+import AdminOffrePage from "./AdminOffrePage"
 import { PAGES } from "@/utils/routes.utils"
-import { useSearchParamsRecord } from "@/utils/useSearchParamsRecord"
 
-export default function UpdateOffre() {
-  const router = useRouter()
-  const { establishment_id, job_id, userId } = useParams() as { establishment_id: string; job_id: string; userId: string }
-  const { raison_sociale } = useSearchParamsRecord()
-  return (
-    <>
-      <Breadcrumb
-        pages={[
-          PAGES.static.backAdminHome,
-          PAGES.dynamic.backAdminRecruteurOffres({ user_id: userId, user_label: raison_sociale }),
-          PAGES.dynamic.offreUpsert({ offerId: job_id, establishment_id, userType: ADMIN }),
-        ]}
-      />
-      <UpsertOffre
-        establishment_id={establishment_id}
-        job_id={job_id}
-        onSuccess={() =>
-          setTimeout(() => {
-            router.push(PAGES.dynamic.successEditionOffre({ establishment_id, userType: "ADMIN", user_id: userId }).getPath())
-          }, 2000)
-        }
-      />
-    </>
-  )
+export async function generateMetadata({ params }: { params: Promise<{ job_id: string; establishment_id: string; userId: string }> }): Promise<Metadata> {
+  const { job_id, establishment_id, userId } = await params
+  return {
+    title: PAGES.dynamic.offreUpsert({ establishment_id, offerId: job_id, userType: ADMIN, userId }).getMetadata().title,
+  }
+}
+
+export default async function Page() {
+  return <AdminOffrePage />
 }
