+++ ui/app/(espace-pro)/espace-pro/(connected)/administration/users/page.tsx
@@ -1,7 +1,12 @@
+import type { Metadata } from "next"
 import { UsersList } from "./UsersList"
 import { AdminLayout } from "@/app/(espace-pro)/espace-pro/(connected)/_components/AdminLayout"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import { PAGES } from "@/utils/routes.utils"
+
+export const metadata: Metadata = {
+  title: PAGES.static.backAdminHome.getMetadata().title,
+}
 
 export default async function AccueilAdministration() {
   return (
