+++ ui/app/(espace-pro)/espace-pro/(connected)/cfa/entreprise/[establishment_id]/creation-offre/page.tsx
@@ -1,19 +1,14 @@
-"use client"
-
-import { useParams, useRouter } from "next/navigation"
-
-import UpsertOffre from "@/app/(espace-pro)/espace-pro/(connected)/_components/UpsertOffre"
-import { Breadcrumb } from "@/app/_components/Breadcrumb"
+import type { Metadata } from "next"
+import CfaCreationOffrePage from "./CfaCreationOffrePage"
 import { PAGES } from "@/utils/routes.utils"
 
-export default function EntrepriseCreationOffrePage() {
-  const router = useRouter()
-  const { establishment_id } = useParams() as { establishment_id: string }
+export async function generateMetadata({ params }: { params: Promise<{ establishment_id: string }> }): Promise<Metadata> {
+  const { establishment_id } = await params
+  return {
+    title: PAGES.dynamic.backCfaEntrepriseCreationOffre(establishment_id).getMetadata().title,
+  }
+}
 
-  return (
-    <>
-      <Breadcrumb pages={[PAGES.static.backCfaHome, PAGES.dynamic.backCfaPageEntreprise(establishment_id), PAGES.dynamic.backCfaEntrepriseCreationOffre(establishment_id)]} />
-      <UpsertOffre establishment_id={establishment_id} onSuccess={() => setTimeout(() => router.push(PAGES.dynamic.backCfaPageEntreprise(establishment_id).getPath()), 2000)} />
-    </>
-  )
+export default async function Page() {
+  return <CfaCreationOffrePage />
 }
