+++ ui/app/(espace-pro)/espace-pro/(connected)/entreprise/page.tsx
@@ -1,7 +1,12 @@
+import type { Metadata } from "next"
 import ListeOffres from "@/app/(espace-pro)/espace-pro/(connected)/_components/ListeOffres"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import { getSession } from "@/utils/getSession"
 import { PAGES } from "@/utils/routes.utils"
+
+export const metadata: Metadata = {
+  title: PAGES.static.backHomeEntreprise.getMetadata().title,
+}
 
 export default async function Page() {
   const { user } = await getSession()
