+++ ui/app/(espace-pro)/espace-pro/(connected)/opco/compte/page.tsx
@@ -1,8 +1,13 @@
+import type { Metadata } from "next"
 import { OPCO } from "shared/constants/recruteur"
 
 import CompteRenderer from "@/app/(espace-pro)/espace-pro/(connected)/_components/CompteRenderer"
 import { Breadcrumb } from "@/app/_components/Breadcrumb"
 import { PAGES } from "@/utils/routes.utils"
+
+export const metadata: Metadata = {
+  title: PAGES.dynamic.compte({ userType: OPCO }).getMetadata().title,
+}
 
 export default function Page() {
   return (
