+++ ui/app/(espace-pro)/espace-pro/(from-mail)/mise-en-relation/[establishment_id]/[job_id]/page.tsx
@@ -1,4 +1,9 @@
+import type { Metadata } from "next"
 import MiseEnRelation from "@/app/(espace-pro)/_components/MiseEnRelation"
+
+export const metadata: Metadata = {
+  title: "Mise en relation avec des organismes de formation - La bonne alternance",
+}
 
 export default async function Page({ params, searchParams }) {
   const { establishment_id, job_id } = await params
