+++ ui/app/(espace-pro)/espace-pro/(from-mail)/offre/[jobType]/[jobId]/cancel/page.tsx
@@ -1,4 +1,9 @@
+import type { Metadata } from "next"
 import { OffreActionPage } from "@/app/(espace-pro)/espace-pro/(from-mail)/offre/[jobType]/[jobId]/OffreActionPage"
+
+export const metadata: Metadata = {
+  title: "Annuler une offre - La bonne alternance",
+}
 
 export default async function Page({ params, searchParams }) {
   const { jobType, jobId } = await params
