+++ ui/app/(espace-pro)/espace-pro/(from-mail)/offre/[jobType]/[jobId]/provided/page.tsx
@@ -1,4 +1,9 @@
+import type { Metadata } from "next"
 import { OffreActionPage } from "@/app/(espace-pro)/espace-pro/(from-mail)/offre/[jobType]/[jobId]/OffreActionPage"
+
+export const metadata: Metadata = {
+  title: "Confirmation de l'offre pourvue - La bonne alternance",
+}
 
 export default async function Page({ params, searchParams }) {
   const { jobType, jobId } = await params
