+++ ui/app/(espace-pro)/espace-pro/(from-mail)/proposition/formulaire/[idFormulaire]/offre/[jobId]/siret/[siretFormateur]/page.tsx
@@ -1,4 +1,9 @@
+import type { Metadata } from "next"
 import { PropositionOffreId } from "./PropositionOffreId"
+
+export const metadata: Metadata = {
+  title: "Proposition d'offre - La bonne alternance",
+}
 
 export default async function Page({ params, searchParams }) {
   const { idFormulaire, jobId, siretFormateur } = (await params) as { idFormulaire: string; jobId: string; siretFormateur: string }
